+++ src/components/atoms/AddNoticeContainer.ts
@@ -4,5 +4,6 @@ import CenterContainer from './CenterContainer';
 export default styled(CenterContainer)`
   width: 100%;
   position: absolute;
-  bottom: 60px;
+  // bottom: 60px; // In case of ServiceMessageLine
+  bottom: 25px;
 `;