+++ src/components/atoms/icons/index.ts
@@ -15,3 +15,4 @@ export { default as ThumbUp } from './ThumbUp';
 export { default as Write } from './Write';
 export { default as Download } from './Download';
 export { default as LinkIcon } from './Link';
+export { default as Plus } from './Plus';