+++ src/webext/createActionsListener.ts
@@ -4,11 +4,11 @@ import { Emit } from 'app/store/types';
 
 const createActionsListener = (emit: Emit) => {
   const messageHandler = createMessageHandler(emit);
-  chrome.runtime.onMessage.addListener(messageHandler);
+  browser.runtime.onMessage.addListener(messageHandler);
   Logger.info('Listening actions ...');
 
   return () => {
-    chrome.runtime.onMessage.removeListener(messageHandler);
+    browser.runtime.onMessage.removeListener(messageHandler);
   };
 };