+++ src/webext/isAuthorizedTab.ts
@@ -2,7 +2,7 @@ import Tab from 'app/lmem/tab';
 import forbiddenTabs from './forbiddenTabs';
 
 export const isTabAuthorizedByPatterns = (patterns: RegExp[]) => (
-  tab: Tab | chrome.tabs.Tab
+  tab: Tab | browser.tabs.Tab
 ) => {
   for (const pattern of patterns) {
     if (!tab.url || pattern.test(tab.url)) return false;