+++ src/webext/types.ts
@@ -1,0 +1,3 @@
+export type InstalledDetails = chrome.runtime.InstalledDetails;
+export type TabActiveInfo = chrome.tabs.TabActiveInfo;
+export type CreateProperties = chrome.tabs.CreateProperties;